恢复全景滑动条对画布缩放的控制，canvas本身尚无缩放功能，所以未测试反向效果；从pro中去掉对canvascontainer组件的编译，但依

+++ src/painttyDesktop/widgets/panoramaslider.h
@@ -13,9 +13,9 @@ class PanoramaSlider : public QWidget
 public:
     explicit PanoramaSlider(QWidget *parent = 0);
 signals:
-    void scaled(qreal scaleFactor);
+    void scaled(int scaleFactor);
 public slots:
-    void setScale(qreal scaleFactor);
+    void setScale(int scaleFactor);
 private:
     QSlider *slider;
     QLineEdit *input;
